refactor: rename dimensionIds -> dimensionId
