+++ src/components/MainSidebar/ProgramDimensionsPanel/useProgramDataDimensions.js
@@ -5,7 +5,7 @@ import {
 } from '@dhis2/analytics'
 import { useDataEngine } from '@dhis2/app-runtime'
 import { useEffect, useReducer, useCallback, useRef, useMemo } from 'react'
-import { extractDimensionIdParts } from '../../../modules/dimensionIds.js'
+import { extractDimensionIdParts } from '../../../modules/dimensionId.js'
 import { DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY } from '../../../modules/userSettings.js'
 import {
     OUTPUT_TYPE_EVENT,
