+++ src/modules/programDimensions.js
@@ -8,7 +8,7 @@ import {
     DIMENSION_TYPE_STATUS,
     DIMENSION_ID_PROGRAM_STATUS,
 } from './dimensionConstants.js'
-import { extractDimensionIdParts } from './dimensionIds.js'
+import { extractDimensionIdParts } from './dimensionId.js'
 import { getDefaultOrgUnitLabel } from './metadata.js'
 import { PROGRAM_TYPE_WITHOUT_REGISTRATION } from './programTypes.js'
 import {
